feat: Introduce product update functionality, standardize error responses, and refactor stock movement reporting.

+++ src/main/java/com/marcosvinirocha/desafio_nexdom/dto/MovimentoEstoqueResponseDTO.java
@@ -3,9 +3,9 @@ package com.marcosvinirocha.desafio_nexdom.dto;
 import com.marcosvinirocha.desafio_nexdom.entity.enums.TipoMovimentacao;
 
 public record MovimentoEstoqueResponseDTO(
-        TipoMovimentacao tipoMovimentacao,
-        Long quantidadeTotalSaida,
-        Integer quantidadeDisponivel
+                TipoMovimentacao tipoMovimentacao,
+                Long quantidadeTotalSaida,
+                Long quantidadeDisponivel
 
 ) {
